# PERF: any/all with axis=1