```
from asv_bench.benchmarks.frame_methods import *
self = Dropna()
self.setup("any", 0)

%timeit self.time_dropna_axis_mixed_dtypes("any", 0)
23.6 ms ± 524 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
10.3 ms ± 190 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```